Added Service and Controller packages.
Modified the computerRepository class: added the function findBybrandContains

Added the ComputerService and ComputerController classes: with related functions to do: 1)addComputer. 2) addComputers which adds a list of computers. 3)  getComputer which returns a computer by Id. 4) getComputerByBrand wwhich returns list of computers by brand. 5)getComputers which returns the list of all computers. 6)deleteComputer which deletes a computer by an Id.