Added Service and Controller packages.
Modified the computerRepository class: added the function findBybrandContains

Added the ComputerService and ComputerController classes: with related functions to do: 1)addFeature. 2)addFeatures which adds a list of features. 3)getFeature which returns a feature by Id. 4)getFeatures which returns the list of all features. 5)deleteFeature which deletes a feature by an Id.